use np.float32 in python script for compatibility with the simulation. Use smaller dt because the CFL criterion is crashing! the implementation of the CFL criterion is defective because it makes use of the minimum velocity instead of the maximum!

+++ apps/visualize.py
@@ -8,7 +8,7 @@ model = sys.argv[2]
 nz = int(sys.argv[3])
 nx = int(sys.argv[4])
 
-float_type = np.float64
+float_type = np.float32
 snapshot_image = np.reshape(np.fromfile(snapshot, dtype=float_type), (nz,nx))
 model_image = np.reshape(np.fromfile(model, dtype=float_type), (nz,nx))
 model_gradient = ndimage.sobel(model_image, axis=0, mode='constant')
